requirements.txt
fastapi==0.111.0
uvicorn==0.23.0
PyJWT==2.8.0
